requirements.txt
python-telegram-bot==13.7
pytz==2024.1
APScheduler==3.6.3 